OpenWhisk/image-resizer/requirements.txt
pillow
requests<=2.18.4
numpy
